requirements.txt
discord.py>=2.4.0
aiohttp>=3.9.0
psycopg2-binary>=2.9.0
statbotics>=2.0.0
